requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dateutil>=2.8.0
matplotlib>=3.7.0
xlsxwriter>=3.1.0
streamlit-searchbox>=0.1.20
